Fix improper rendering of underscored variables
The underscore variable __nativeMethod__ (two underscores at start + end) was not being properly rendered. This is possibly because underscores have a significance of their own in Markdown. So I moved the content from a TABLE to a CODEBLOCK, which suppresses Markdown formatting. Now the rendering is correct. No change in content.